test-requirements.txt
pytest
pytest-cov
pytest-mock
requests
pytest-flask
pytest-timeout
fakeredis
selenium
webdriver-manager